Move functions rewrite* and replace* to pyk.subst
